Created a Database handler.
This commit closes in parts the issue #31
Implemented a Database handler and wrote a test, wich should fail untill
database is created with a test subject with id = 1.

The Database handler uses preparedstatments.
Database.js exports a number of suggested skeleton functions to be
implemented if @Admin agrees with its usefullness.

The query function in the Database.js is a bit complicated and could use
a bit of improvment, but I have faild in finding a more suitable option.